requirements.txt
attr==0.3.2
ConfigParser==7.1.0
cryptography==43.0.1
Cython==3.0.11
dl==0.1.0
docutils==0.21.2
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.5
Jinja2==3.1.4
jnius==1.1.0
keyring==25.4.0
lockfile==0.12.2
numpy==2.1.1
Pillow==9.5.0
Pillow==10.4.0
protobuf==5.28.2
pyOpenSSL==24.2.1
pywin32==306
railroad==0.5.0
rembg==2.0.59
Sphinx==8.0.2
thread==2.0.5
tornado==6.4.1
trove_classifiers==2024.9.12
truststore==0.9.2
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
